requirements.txt
pip==19.0.3
numpy==1.18.1
pandas==0.24
pydub==0.23.1
protobuf==3.11.3
tensorflow==1.15.0
matplotlib
scipy==1.4.1
pyyaml==5.3
scikit-learn==0.22.2
